Moved modules to separate files, fixed inaccurate comments and spacing on curly braces
