+++ lib/devise/multi_email/models/validatable.rb
@@ -78,7 +78,7 @@ module Devise
                        :validates_confirmation_of, :validates_length_of].freeze
 
         def assert_validations_api!(base) #:nodoc:
-          unavailable_validations = VALIDATIONS.select { |v| !base.respond_to?(v) }
+          unavailable_validations = VALIDATIONS.select{ |v| !base.respond_to?(v) }
 
           unless unavailable_validations.empty?
             raise "Could not use :validatable module since #{base} does not respond " <<
